requirements.txt
PyYAML==6.0.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.27.2
rich==13.9.2
SQLAlchemy==2.0.36
pydantic==2.9.2
feedparser==6.0.11
scrapy==2.11.2
boto3==1.35.28
pdfminer.six==20231228
psycopg2-binary==2.9.9
paperscraper>=0.3.0
habanero>=1.2.0
pyalex>=1.0.0
semanticscholar>=2.0.0
